re-organize testing of base catalog class and its subclasses
